+++ domestic/static/javascript/showMore.js
@@ -1,9 +1,18 @@
 const handleToggle = (element) => {
   const text = element.previousElementSibling
-  return (
-    text.classList.toggle("details-text-open")
-  )
+  const summary = element.querySelector('summary')
+  const summaryText = summary.innerText
+
+  if (summaryText === 'Show more') {
+    summary.innerText = 'Show less'
+  } else {
+    summary.innerText = 'Show more'
+  }
+
+  return text.classList.toggle('details-text-open')
 }
 
 const detailsElements = document.querySelectorAll('[data-show-more]')
-detailsElements.forEach((element) => element.addEventListener('toggle', () => handleToggle(element)))
+detailsElements.forEach((element) =>
+  element.addEventListener('toggle', () => handleToggle(element))
+)