+++ export_academy/forms.py
@@ -143,7 +143,7 @@ class EventAdminModelForm(WagtailAdminModelForm):
         field_value = self.cleaned_data.get(field)
         if event_format == self.Meta.model.IN_PERSON:
             if not field_value:
-                self._errors[field] = self.error_class(["In-person event requires this field."])
+                self._errors[field] = self.error_class(['In-person event requires this field.'])
 
         return field_value