+++ international_online_offer/views.py
@@ -3,6 +3,12 @@ from django.views.generic import TemplateView
 from django.views.generic.edit import FormView
 
 from international_online_offer import forms
+from international_online_offer.models import (
+    TriageData,
+    UserData,
+    get_triage_data,
+    get_user_data,
+)
 
 LOW_VALUE_INVESTOR_CONTACT_FORM_MESSAGE = 'Complete the contact form to keep up to date with our personalised service.'
 HIGH_VALUE_INVESTOR_CONTACT_FORM_MESSAGE = """Your business qualifies for 1 to 1 support from specialist UK government
@@ -30,11 +36,20 @@ class IOOSector(FormView):
         )
 
     def get_initial(self):
-        inital_sector = self.request.session.get('sector')
-        return {'sector': inital_sector}
+        if self.request.user.is_authenticated:
+            triage_data = get_triage_data(self.request.user.hashed_uuid)
+            if triage_data:
+                return {'sector': triage_data.sector}
+
+        return {'sector': self.request.session.get('sector')}
 
     def form_valid(self, form):
-        self.request.session['sector'] = form.cleaned_data['sector']
+        if self.request.user.is_authenticated:
+            TriageData.objects.update_or_create(
+                hashed_uuid=self.request.user.hashed_uuid, defaults={'sector': form.cleaned_data['sector']}
+            )
+        else:
+            self.request.session['sector'] = form.cleaned_data['sector']
         return super().form_valid(form)
 
 
@@ -54,13 +69,22 @@ class IOOIntent(FormView):
         )
 
     def get_initial(self):
-        inital_intent = self.request.session.get('intent')
-        inital_intent_other = self.request.session.get('intent_other')
-        return {'intent': inital_intent, 'intent_other': inital_intent_other}
+        if self.request.user.is_authenticated:
+            triage_data = get_triage_data(self.request.user.hashed_uuid)
+            if triage_data:
+                return {'intent': triage_data.intent, 'intent_other': triage_data.intent_other}
+
+        return {'intent': self.request.session.get('intent'), 'intent_other': self.request.session.get('intent_other')}
 
     def form_valid(self, form):
-        self.request.session['intent'] = form.cleaned_data['intent']
-        self.request.session['intent_other'] = form.cleaned_data['intent_other']
+        if self.request.user.is_authenticated:
+            TriageData.objects.update_or_create(
+                hashed_uuid=self.request.user.hashed_uuid,
+                defaults={'intent': form.cleaned_data['intent'], 'intent_other': form.cleaned_data['intent_other']},
+            )
+        else:
+            self.request.session['intent'] = form.cleaned_data['intent']
+            self.request.session['intent_other'] = form.cleaned_data['intent_other']
         return super().form_valid(form)
 
 
@@ -80,13 +104,28 @@ class IOOLocation(FormView):
         )
 
     def get_initial(self):
-        inital_location = self.request.session.get('location')
-        inital_location_none = self.request.session.get('location_none')
-        return {'location': inital_location, 'location_none': inital_location_none}
+        if self.request.user.is_authenticated:
+            triage_data = get_triage_data(self.request.user.hashed_uuid)
+            if triage_data:
+                return {'location': triage_data.location, 'location_none': triage_data.location_none}
+
+        return {
+            'location': self.request.session.get('location'),
+            'location_none': self.request.session.get('location_none'),
+        }
 
     def form_valid(self, form):
-        self.request.session['location'] = form.cleaned_data['location']
-        self.request.session['location_none'] = form.cleaned_data['location_none']
+        if self.request.user.is_authenticated:
+            TriageData.objects.update_or_create(
+                hashed_uuid=self.request.user.hashed_uuid,
+                defaults={
+                    'location': form.cleaned_data['location'],
+                    'location_none': form.cleaned_data['location_none'],
+                },
+            )
+        else:
+            self.request.session['location'] = form.cleaned_data['location']
+            self.request.session['location_none'] = form.cleaned_data['location_none']
         return super().form_valid(form)
 
 
@@ -106,11 +145,20 @@ class IOOHiring(FormView):
         )
 
     def get_initial(self):
-        inital_hiring = self.request.session.get('hiring')
-        return {'hiring': inital_hiring}
+        if self.request.user.is_authenticated:
+            triage_data = get_triage_data(self.request.user.hashed_uuid)
+            if triage_data:
+                return {'hiring': triage_data.hiring}
+
+        return {'hiring': self.request.session.get('hiring')}
 
     def form_valid(self, form):
-        self.request.session['hiring'] = form.cleaned_data['hiring']
+        if self.request.user.is_authenticated:
+            TriageData.objects.update_or_create(
+                hashed_uuid=self.request.user.hashed_uuid, defaults={'hiring': form.cleaned_data['hiring']}
+            )
+        else:
+            self.request.session['hiring'] = form.cleaned_data['hiring']
         return super().form_valid(form)
 
 
@@ -130,13 +178,22 @@ class IOOSpend(FormView):
         )
 
     def get_initial(self):
-        inital_spend = self.request.session.get('spend')
-        inital_spend_other = self.request.session.get('spend_other')
-        return {'spend': inital_spend, 'spend_other': inital_spend_other}
+        if self.request.user.is_authenticated:
+            triage_data = get_triage_data(self.request.user.hashed_uuid)
+            if triage_data:
+                return {'spend': triage_data.spend, 'spend_other': triage_data.spend_other}
+
+        return {'spend': self.request.session.get('spend'), 'spend_other': self.request.session.get('spend_other')}
 
     def form_valid(self, form):
-        self.request.session['spend'] = form.cleaned_data['spend']
-        self.request.session['spend_other'] = form.cleaned_data['spend_other']
+        if self.request.user.is_authenticated:
+            TriageData.objects.update_or_create(
+                hashed_uuid=self.request.user.hashed_uuid,
+                defaults={'spend': form.cleaned_data['spend'], 'spend_other': form.cleaned_data['spend_other']},
+            )
+        else:
+            self.request.session['spend'] = form.cleaned_data['spend']
+            self.request.session['spend_other'] = form.cleaned_data['spend_other']
         return super().form_valid(form)
 
 
@@ -151,6 +208,52 @@ class IOOContact(FormView):
             complete_contact_form_message=LOW_VALUE_INVESTOR_CONTACT_FORM_MESSAGE,
             back_url='/international/expand-your-business-in-the-uk/guide/',
         )
+
+    def get_initial(self):
+        if self.request.user.is_authenticated:
+            user_data = get_user_data(self.request.user.hashed_uuid)
+            if user_data:
+                return {
+                    'company_name': user_data.company_name,
+                    'company_location': user_data.company_location,
+                    'full_name': user_data.full_name,
+                    'role': user_data.role,
+                    'email': user_data.email,
+                    'telephone_number': user_data.telephone_number,
+                    'agree_terms': user_data.agree_terms,
+                    'agree_info_email': user_data.agree_info_email,
+                    'agree_info_telephone': user_data.agree_info_telephone,
+                }
+
+        return {
+            'company_name': self.request.session.get('company_name'),
+            'company_location': self.request.session.get('company_location'),
+            'full_name': self.request.session.get('full_name'),
+            'role': self.request.session.get('role'),
+            'email': self.request.session.get('email'),
+            'telephone_number': self.request.session.get('telephone_number'),
+            'agree_terms': self.request.session.get('agree_terms'),
+            'agree_info_email': self.request.session.get('agree_info_email'),
+            'agree_info_telephone': self.request.session.get('agree_info_telephone'),
+        }
+
+    def form_valid(self, form):
+        if self.request.user.is_authenticated:
+            UserData.objects.update_or_create(
+                hashed_uuid=self.request.user.hashed_uuid,
+                defaults=form.cleaned_data,
+            )
+        else:
+            self.request.session['company_name'] = form.cleaned_data['company_name']
+            self.request.session['company_location'] = form.cleaned_data['company_location']
+            self.request.session['full_name'] = form.cleaned_data['full_name']
+            self.request.session['role'] = form.cleaned_data['role']
+            self.request.session['email'] = form.cleaned_data['email']
+            self.request.session['telephone_number'] = form.cleaned_data['telephone_number']
+            self.request.session['agree_terms'] = form.cleaned_data['agree_terms']
+            self.request.session['agree_info_email'] = form.cleaned_data['agree_info_email']
+            self.request.session['agree_info_telephone'] = form.cleaned_data['agree_info_telephone']
+        return super().form_valid(form)
 
 
 class IOOLogin(FormView):