+++ tests/unit/international_online_offer/test_views.py
@@ -1,6 +1,8 @@
 import pytest
 from django.urls import reverse
 
+from international_online_offer.models import TriageData
+
 
 @pytest.mark.django_db
 def test_ioo_index(client, settings):
@@ -13,9 +15,40 @@ def test_ioo_index(client, settings):
 @pytest.mark.django_db
 def test_ioo_sector(client, settings):
     settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    response = client.get(reverse('international_online_offer:sector'))
+    assert response.status_code == 200
+
+
+@pytest.mark.django_db
+def test_ioo_sector_initial(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    TriageData.objects.update_or_create(
+        hashed_uuid='123',
+        defaults={'sector': 'sector'},
+    )
     url = reverse('international_online_offer:sector')
+    user.hashed_uuid = '123'
+    client.force_login(user)
     response = client.get(url)
     assert response.status_code == 200
+
+
+@pytest.mark.django_db
+def test_ioo_sector_form_valid_saves_to_db(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:sector')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.post(url, {'sector': 'Food and Drink'})
+    assert response.status_code == 302
+
+
+@pytest.mark.django_db
+def test_ioo_sector_form_valid_saves_to_session(client, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:sector')
+    response = client.post(url, {'sector': 'Food and Drink'})
+    assert response.status_code == 302
 
 
 @pytest.mark.django_db
@@ -32,6 +65,38 @@ def test_ioo_intent(client, settings):
     url = reverse('international_online_offer:intent')
     response = client.get(url)
     assert response.status_code == 200
+
+
+@pytest.mark.django_db
+def test_ioo_intent_initial(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    TriageData.objects.update_or_create(
+        hashed_uuid='123',
+        defaults={'intent': ['Set up new premises'], 'intent_other': ''},
+    )
+    url = reverse('international_online_offer:intent')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.get(url)
+    assert response.status_code == 200
+
+
+@pytest.mark.django_db
+def test_ioo_intent_form_valid_saves_to_db(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:intent')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.post(url, {'intent': 'Set up new premises', 'intent_other': ''})
+    assert response.status_code == 302
+
+
+@pytest.mark.django_db
+def test_ioo_intent_form_valid_saves_to_session(client, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:intent')
+    response = client.post(url, {'intent': 'Set up new premises', 'intent_other': ''})
+    assert response.status_code == 302
 
 
 @pytest.mark.django_db
@@ -52,6 +117,38 @@ def test_ioo_location(client, settings):
 
 
 @pytest.mark.django_db
+def test_ioo_location_initial(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    TriageData.objects.update_or_create(
+        hashed_uuid='123',
+        defaults={'location': 'location'},
+    )
+    url = reverse('international_online_offer:location')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.get(url)
+    assert response.status_code == 200
+
+
+@pytest.mark.django_db
+def test_ioo_location_form_valid_saves_to_db(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:location')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.post(url, {'location': 'Wales'})
+    assert response.status_code == 302
+
+
+@pytest.mark.django_db
+def test_ioo_location_form_valid_saves_to_session(client, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:location')
+    response = client.post(url, {'location': 'Wales'})
+    assert response.status_code == 302
+
+
+@pytest.mark.django_db
 def test_triage_location_session(client, settings):
     settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
     url = reverse('international_online_offer:location')
@@ -69,6 +166,38 @@ def test_ioo_hiring(client, settings):
 
 
 @pytest.mark.django_db
+def test_ioo_hiring_initial(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    TriageData.objects.update_or_create(
+        hashed_uuid='123',
+        defaults={'hiring': 'hiring'},
+    )
+    url = reverse('international_online_offer:hiring')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.get(url)
+    assert response.status_code == 200
+
+
+@pytest.mark.django_db
+def test_ioo_hiring_form_valid_saves_to_db(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:hiring')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.post(url, {'hiring': '1-10'})
+    assert response.status_code == 302
+
+
+@pytest.mark.django_db
+def test_ioo_hiring_form_valid_saves_to_session(client, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:hiring')
+    response = client.post(url, {'hiring': '1-10'})
+    assert response.status_code == 302
+
+
+@pytest.mark.django_db
 def test_triage_hiring_session(client, settings):
     settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
     url = reverse('international_online_offer:hiring')
@@ -82,6 +211,38 @@ def test_ioo_spend(client, settings):
     url = reverse('international_online_offer:spend')
     response = client.get(url)
     assert response.status_code == 200
+
+
+@pytest.mark.django_db
+def test_ioo_spend_initial(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    TriageData.objects.update_or_create(
+        hashed_uuid='123',
+        defaults={'spend': 'spend'},
+    )
+    url = reverse('international_online_offer:spend')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.get(url)
+    assert response.status_code == 200
+
+
+@pytest.mark.django_db
+def test_ioo_spend_form_valid_saves_to_db(client, user, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:spend')
+    user.hashed_uuid = '123'
+    client.force_login(user)
+    response = client.post(url, {'spend': '10000-500000'})
+    assert response.status_code == 302
+
+
+@pytest.mark.django_db
+def test_ioo_spend_form_valid_saves_to_session(client, settings):
+    settings.FEATURE_INTERNATIONAL_ONLINE_OFFER = True
+    url = reverse('international_online_offer:spend')
+    response = client.post(url, {'spend': '10000-500000'})
+    assert response.status_code == 302
 
 
 @pytest.mark.django_db